requirements.txt
astroid==2.5.1
cachelib==0.1.1
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pylint==2.7.2
python-dotenv==0.15.0
redis==3.5.3
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
Werkzeug==1.0.1
wrapt==1.12.1
